Combined table and datacard in one component + responsive demo
