+++ src/components/Table/DataCard/style.tsx
@@ -2,7 +2,7 @@ import _R from 'react';
 import { StyledComponentClass as _S } from 'styled-components';
 import _T from '../../../types/ThemeType';
 import styled, { withProps } from '../../../utility/_styled';
-import StyledCell from '../Cell/style';
+import StyledCell from '../DataCardCell/style';
 
 type StyledRowProps = {
     ref?: HTMLElement;
