+++ src/components/Table/DataCardCell/style.tsx
