+++ src/components/Table/tableStyle.tsx
